:hammer: Migrate to Spring Boot, remove Mysql and add postgresql :construction:
