+++ README.md
@@ -1,0 +1,4 @@
+# Broj Service
+
+### Getting Started
+- `mvn clean package`
