--- src/main/java/com/broj/config/RootConfig.java
@@ -1,12 +1,0 @@
-package com.broj.config;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * @author Muztaba Hasanat
- * @author Rimon Mostafiz
- */
-@Configuration
-@ComponentScan(basePackages={"com.broj"})
-public class RootConfig {}
