--- src/main/java/com/broj/config/WebAppInit.java
@@ -1,27 +1,0 @@
-package com.broj.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
-
-/**
- * @author Muztaba Hasanat
- * @author Rimon Mostafiz
- */
-@Configuration
-public class WebAppInit extends AbstractAnnotationConfigDispatcherServletInitializer {
-    @Override
-    protected Class<?>[] getRootConfigClasses() {
-        return new Class<?>[]{RootConfig.class};
-    }
-
-    @Override
-    protected Class<?>[] getServletConfigClasses() {
-        return new Class<?>[]{WebConfig.class};
-    }
-
-    @Override
-    protected String[] getServletMappings() {
-        return new String[]{"/"};
-    }
-
-}
