--- src/main/java/com/broj/service/Compile.java
@@ -1,14 +1,0 @@
-package com.broj.service;
-
-import com.broj.model.CompileRequest;
-import com.broj.model.CompileResponse;
-
-/**
- * @author Muztaba Hasanat
- * @author Rimon Mostafiz
- */
-public interface Compile {
-
-    CompileResponse compile(CompileRequest request);
-
-}
