+++ src/main/resources/application.properties
@@ -1,0 +1,17 @@
+server.servlet.context-path=/broj-service
+server.port=8080
+
+#LOGGING
+logging.level.root=info
+logging.level.com.rimonmostafiz=debug
+
+#DATASOURCE PROPERTIES
+spring.datasource.url=jdbc:postgresql://postgresqldb:5432/broj-db
+spring.datasource.username=postgres
+spring.datasource.password=postgres
+spring.datasource.driver-class-name=org.postgresql.Driver
+
+spring.jpa.show-sql=true
+spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto=create
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL95Dialect
