+++ src/test/java/com/rimonmostafiz/broj/BrojServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.rimonmostafiz.broj;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BrojServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
